// frontend2/src/api/orval/client/model/canvasEngineType.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Military Journal API
 * OpenAPI spec version: 1.0.0
 */

/**
 * Engine type for the canvas editor.
 */
export type CanvasEngineType = typeof CanvasEngineType[keyof typeof CanvasEngineType];


export const CanvasEngineType = {
  konva: 'konva',
  fabric: 'fabric',
} as const;